Move majority of NaNo code to its own file
The parts that are left in kalpana.py3 and terminal.py have been
modified so that all the interesting stuff happens in nano.py.
So far, this is very experimental because aaargh objects are my Nemesis.
Sligthly checked for bugs, meaning that nothing actually explodes when the
NaNo thingy is activated.
Still left to do: find out what the code does, and why, and if it can be
better at it.