The n command can now set the filename
This is basically the same functionality as starting kalpana with an arg
that is not an existing file. The filename is set to the argument and
the next time you save the file will be created with that name and path.
(Obviously this doesn't hold if you save with s and another filename as
an argument, duh.)
Fixes the last of #76.

Note that this requires the latest changes made in libsyntyche.