Improve formatting performance a bazillion times
It now only re-evaluates the whole document on a few occasions, which
eventually should be possible to get rid of as well (possibly with an
extra "force update/redraw" command) which means that writing a lot of
stuff doesn't lag like shit anymore.

On the downside, right now the word counting is not updated enough.